FunctionApp/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
adal
requests
azure-common==1.1.18
azure-mgmt-containerinstance==1.4.1
azure-mgmt-resource==2.1.0
